+++ react/Product.js
@@ -146,6 +146,7 @@ export const parseToJsonLD = ({
   product,
   selectedItem,
   currency,
+  disableOffers,
   decimals,
   pricesWithTax,
 }) => {
@@ -175,7 +176,7 @@ export const parseToJsonLD = ({
     mpn: product.productId,
     sku: selectedItem && selectedItem.itemId,
     category: getCategoryName(product),
-    offers,
+    offers: disableOffers ? null : offers,
   }
 
   return productLD
@@ -186,12 +187,13 @@ function StructuredData({ product, selectedItem }) {
     culture: { currency },
   } = useRuntime()
 
-  const { decimals, pricesWithTax } = useAppSettings()
+  const { disableOffers, decimals, pricesWithTax } = useAppSettings()
 
   const productLD = parseToJsonLD({
     product,
     selectedItem,
     currency,
+    disableOffers,
     decimals,
     pricesWithTax,
   })