+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Add new `settingsSchema` option `disableOffers` to fill `offers` as null when enabled
+
 ## [0.10.0] - 2023-02-15
 
 ### Changed